Move PERMANOVA helpers into evaluation.py
- Inline adonis2-compatible pseudo-F and permutation test next to other eval helpers
- Remove separate _permanova module; tests import from evaluation
